refactor(mcp): promote transport security to its own concern
Move build_transport_security out of core/ into a package-root transport_security module, a peer of auth/ rather than a module under the composition tier. It imports only auth (ResourceServer) and settings and is consumed by the runtime, so it slots into the layering contract between core and auth. This is the incremental app-level step toward the target architecture, where core/ dissolves into named concerns.
